+++ src/Streamly/Internal/Data/Parser.hs
@@ -231,7 +231,7 @@ fromFold = K.toParserK . D.fromFold
 -- Terminating but not failing folds
 -------------------------------------------------------------------------------
 --
--- This is the dual of stream "yield".
+-- This is the dual of stream "value".
 --
 -- | A parser that always yields a pure value without consuming any input.
 --