+++ src/Streamly/Prelude.hs
@@ -107,7 +107,7 @@ module Streamly.Prelude
 
     -- ** From Values
     -- | Generate a monadic stream from a seed value or values.
-    , yield
+    , value
     , yieldM
     , repeat
     , repeatM
@@ -582,12 +582,12 @@ module Streamly.Prelude
     -- One dimension loops are just a special case of nested loops.  For
     -- example, 'concatMap' can degenerate to a simple map operation:
     --
-    -- > map f m = S.concatMap (\x -> S.yield (f x)) m
+    -- > map f m = S.concatMap (\x -> S.value (f x)) m
     --
     -- Similarly, 'concatMap' can perform filtering by mapping an element to a
     -- 'nil' stream:
     --
-    -- > filter p m = S.concatMap (\x -> if p x then S.yield x else S.nil) m
+    -- > filter p m = S.concatMap (\x -> if p x then S.value x else S.nil) m
     --
 
     , concatMapWith
@@ -633,6 +633,7 @@ module Streamly.Prelude
 
     -- * Deprecated
     , once
+    , yield
     , each
     , scanx
     , foldx